todo/requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
colorama==0.4.4
fastapi==0.75.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
Jinja2==3.1.1
MarkupSafe==2.1.1
pydantic==1.9.0
python-dotenv==0.20.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing-extensions==4.2.0
uvicorn==0.17.6
